+++ hqlib/metric/team/spirit.py
@@ -18,24 +18,21 @@ from __future__ import absolute_import
 import datetime
 
 from ... import metric_source
-from ...domain import Metric
+from ...domain import Metric, LowerIsBetterMetric
 
 
 class TeamSpirit(Metric):
     """ Metric for measuring the spirit of a specific team. The team simply picks a smiley. """
 
-    name = 'Team stemming'
+    name = 'Teamstemming'
     unit = ''
     norm_template = 'De stemming wordt door het team zelf bepaald door het kiezen van een smiley. ' \
-        'De norm hierbij is een tevreden team, neutraal is geel, ontevreden is rood. ' \
-        'Als de meting ouder is dan {old_age} dagen is de status geel, ouder dan {max_old_age} dagen is rood.'
+        'De norm hierbij is een tevreden team, neutraal is geel, ontevreden is rood.'
     template = 'De stemming van team {name} was {value} op {date}.'
     target_value = ':-)'
     perfect_value = ':-)'
     low_target_value = ':-('
     numerical_value_map = {':-(': 0, ':-|': 1, ':-)': 2}
-    old_age = datetime.timedelta(days=21)
-    max_old_age = 2 * old_age
     metric_source_classes = (metric_source.TeamSpirit,)
 
     def value(self):
@@ -49,21 +46,35 @@ class TeamSpirit(Metric):
         return min(values), max(values)
 
     def _needs_immediate_action(self):
-        # First check whether the metric needs immediate action because it was measured too long ago.
-        # If not, check whether the spirit is too low.
-        return True if super(TeamSpirit, self)._needs_immediate_action() else self.value() == self.low_target()
+        return self.value() == self.low_target()
 
     def _is_value_better_than(self, target):
         return self.numerical_value() > self.numerical_value_map.get(target, -1)
 
     def _is_below_target(self):
-        # First check whether the metric needs action because it was measured too long ago.
-        # If not, check whether the spirit is low.
-        return True if super(TeamSpirit, self)._is_below_target() else \
-            self.numerical_value() < max(self.numerical_value_map.values())
-
-    def _date(self):
-        return self._metric_source.datetime(self._metric_source_id)
+        return self.numerical_value() < max(self.numerical_value_map.values())
 
     def _missing(self):
         return not self._metric_source.team_spirit(self._metric_source_id)
+
+
+class TeamSpiritAge(LowerIsBetterMetric):
+    """ Metric for measuring the time since the team spirit was last updated by the team. """
+    name = 'Teamstemmingleeftijd'
+    unit = 'dagen'
+    norm_template = 'De teamstemming is maximaal {target} {unit} geleden bepaald. ' \
+                    'Langer dan {low_target} {unit} geleden is rood.'
+    template = 'De stemming van team {name} is {value} {unit} geleden bepaald.'
+    target_value = 21
+    low_target_value = 42
+    metric_source_classes = (metric_source.TeamSpirit,)
+
+    def value(self):
+        return -1 if self._missing() else \
+            (datetime.datetime.now() - self._metric_source.datetime(self._metric_source_id)).days
+
+    def _missing(self):
+        if self._metric_source_id:
+            return self._metric_source.datetime(self._metric_source_id) in (None, datetime.datetime.min)
+        else:
+            return True